[TASK] Fluid: Register "core" namespace globally
The Fluid namespace "core" is now available globally, making
it obsolete to register the namespace manually in each file
any core viewhelper is necessary.

This is also a precursor for integrating Frontend Editing content
editable ViewHelpers in all places of TYPO3 Frontend when
fluid templates are in use.

Resolves: #79817
Releases: master
Change-Id: I7691107fa7c02e71b848cd8d0bb0eecdae5a8171
Reviewed-on: https://review.typo3.org/51688
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Mattias Nilsson <tollepjaer@gmail.com>
Tested-by: Mattias Nilsson <tollepjaer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>