[FEATURE] Introduce device presets and redesign the view module
The view module was redesigned to provide a more modern and streamlined look
and feel across the existing backend. With the introduction of named and
categorized device presets we enable users to get a better idea of how the
page will look like on a specific device. For even more easy testing, the
orientation can now be changed without selecting a different device.

Releases: master
Resolves: #77576
Change-Id: Ie430db2a39a9bf6931ebde8109e62c12d9bae09b
Reviewed-on: https://review.typo3.org/53061
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>