[FEATURE] Replace hooks with events for Preview URI generation
The BackendUtility method "getPreviewUrl" along with
its hooks and the helper method "ADMCMD_previewCmds"
has been marked as deprecated in favor of having
everything in PreviewUriBuilder.

This class now contains two new Events for manipulating
the generated preview URI.

In addition, the unused "viewUrl" property of
EditDocumentController is removed.

The language parameter is now additionally
encapsulated in a separate property.

Resolves: #97544
Related: #91123
Releases: main
Change-Id: I5d6fbc293cd5bbfa1d4b522eb2f02bb6c763bb1a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74499
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>